tests/performance/ecommerce_hybrid_search/common/requirements.txt
docker
elasticsearch
pyvespa
tqdm
requests
zstandard